Count enc

+++ source/aogmaneo/Encoder.cpp
@@ -185,7 +185,7 @@ void Encoder::initRandom(
         vl.weights.resize(numHiddenCells * area * vld.size.z);
 
         for (int i = 0; i < vl.weights.size(); i++)
-            vl.weights[i] = static_cast<int>(rand() % 5);
+            vl.weights[i] = rand() % 5;
     }
 
     hiddenCIs = IntBuffer(numHiddenColumns, 0);
